Optimized Memory Usage
Learned about bytes, sbytes, shorts, ushorts and other structs, that work like integers, but with less precision
(but they are more memory efficient)

+++ Assets/C# Scripts/BiomeSettingsSO.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Biome", menuName = "WorldSettings/Biome")]
 public class BiomeSettingsSO : ScriptableObject
 {
-    public int maxChunkHeight;
+    public byte maxChunkHeight;
     public float scale;
     public int octaves;
     public float persistence;
